+++ .babelrc
@@ -3,5 +3,8 @@
     "es2015",
     "stage-0",
     "react"
+  ],
+  "plugins": [
+    "react-hot-loader/babel"
   ]
 }
