+++ .npmrc
@@ -1,0 +1,1 @@
+registry=https://registry.npm.taobao.org
