+++ src/Breadcrumb/Breadcrumb.js
@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import classNames from 'classnames';
 import {ItemDropDown, MenuItem} from '../Menu';
