+++ src/Modal/layer.js
@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import Modal from './Modal';
 import Button from '../Button/Button';
