+++ src/Tree/Tree.js
@@ -1,4 +1,5 @@
-import React, {PropTypes} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import {loopChildren, filterParentPosition, handleCheck, getCheck, isInclude, arrayEqual} from '../utils/treeUtils'
 import TreeNode from './TreeNode';
