+++ src/internal/Overlay.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ScrollLock from '../internal/ScrollLock';
 
 function getStyles(props){
