+++ src/internal/ScrollLock.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 let lockingCounter = 0;
 let originalBodyOverflow = null;
