+++ src/utils/shouldPureComponentUpdate.js
@@ -1,0 +1,32 @@
+//https://github.com/gaearon/react-pure-render
+export const shallowEqual = (objA, objB) => {
+    if (objA === objB) {
+        return true;
+    }
+
+    if (typeof objA !== 'object' || objA === null ||
+        typeof objB !== 'object' || objB === null) {
+        return false;
+    }
+
+    var keysA = Object.keys(objA);
+    var keysB = Object.keys(objB);
+
+    if (keysA.length !== keysB.length) {
+        return false;
+    }
+
+    // Test for A's keys different from B.
+    var bHasOwnProperty = Object.prototype.hasOwnProperty.bind(objB);
+    for (var i = 0; i < keysA.length; i++) {
+        if (!bHasOwnProperty(keysA[i]) || objA[keysA[i]] !== objB[keysA[i]]) {
+            return false;
+        }
+    }
+    return true;
+};
+
+export default function shouldPureComponentUpdate(nextProps, nextState) {
+    return !shallowEqual(this.props, nextProps) ||
+        !shallowEqual(this.state, nextState);
+}
